pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "en-core-web-sm",
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.7",
    "langsmith>=0.6.5",
    "networkx>=3.6.1",
    "pinecone-client>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
